Finished Prototype UI
-Bars keep shape as they decrease
-Shotgun now takes ammo to use
-Garrett's UI has values for the bars. Wondering which one looks better.

+++ Project_Thieves/Assets/Scripts/PlayerScripts/GarrettScripts/Weapon_Gun.cs
@@ -12,7 +12,7 @@ public class Weapon_Gun : MonoBehaviour
     public ShotType shotType;
 
     public Bullet bullet;      //Which bullet prefab will this weapon fire
-    public float ammoCost;          //How much ammo does this weapon cost
+    public int ammoCost;          //How much ammo does this weapon cost
 
     [Range(0f, 360f)]
     public float spreadAngle;       //How much spread on the shot
